fastapi_backend/requirements.txt
api==0.0.7
fastapi==0.109.0
keras==2.15.0
numpy==1.23.5
Pillow==10.2.0
pydantic==1.10.8
tensorflow==2.15.0.post1
uvicorn==0.27.0
